Move commit message generation to commitinfo pkg
